[REFACTOR] Unify source image and UV pose calculations
